refactor: move descriptor utility functions to a separate file
Issue: BTC-1348
